Added GFTabBarController
- Removed tab bar functions in SceneDelegate to a custom GFTabBarController
