+++ GHFollowers/Custom Views/TabBarControllers/GFTabBarController.swift
@@ -1,0 +1,34 @@
+//
+//  GFTabBarController.swift
+//  GHFollowers
+//
+//  Created by Brandon Tyler on 2/18/20.
+//  Copyright © 2020 Brandon Maynard. All rights reserved.
+//
+
+import UIKit
+
+class GFTabBarController: UITabBarController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        UITabBar.appearance().tintColor = .systemGreen
+        self.viewControllers = [createSearchNC(), createFavoritesNC()]
+    }
+    
+    func createSearchNC() -> UINavigationController {
+        let searchVC = SearchVC()
+        searchVC.title = "Search"
+        searchVC.tabBarItem = UITabBarItem(tabBarSystemItem: .search, tag: 0)
+        
+        return UINavigationController(rootViewController: searchVC)
+    }
+    
+    func createFavoritesNC() -> UINavigationController {
+        let favoritesListVC = FavoritesListVC()
+        favoritesListVC.title = "Favorites"
+        favoritesListVC.tabBarItem = UITabBarItem(tabBarSystemItem: .favorites, tag: 1)
+        
+        return UINavigationController(rootViewController: favoritesListVC)
+    }
+}
